refactor(i18n): run the unit-literal scan as a gate, not a vitest test
The scan parses every in-scope file with the TypeScript compiler, so its cost
scales with the size of the repo rather than with what a branch changed. Inside
vitest that cost is multiplied by v8 coverage instrumentation and by every
sibling worker. Measured on one 6-core box, same commit, only --maxWorkers
varying:

  standalone gate ....  2.0 s
   2 workers .........  11.2 s   5.6x
   4 workers .........  13.7 s   6.8x   <- GitHub CI shape, 15 s budget
   6 workers .........  16.6 s   8.2x
   8 workers .........  20.3 s  10.1x
  12 workers .........  28.6 s  14.2x

The suite's 15 s per-test budget, sized for tests that `await import(...)`, is
already marginal at the CI worker count and fails outright on a developer
machine. Raising it only moves the number, because the cost tracks worker count.
Throttling workers is worse: 12 -> 2 cuts the scan 60% but costs the whole suite
82% more wall clock (512 s -> 929 s). Even at 2 workers the scan pays a 5.6x tax
it cannot escape, because v8 coverage is process-wide. There is no worker count
at which this belongs in the suite.

So it moves out, following the pattern this repo already uses for repo-scale
gates (check-source-strings.mjs, check-i18n-strings.mjs, docs-lint.sh):

- scripts/lib/unit-literals.mjs    the matcher, shared by the gate and the tests
- scripts/check-unit-literals.mjs  the gate: ceiling + two diff-scoped checks
- one line in the i18n gate table, which CI already runs; no workflow change
- unitLiterals.test.ts keeps the four matcher tests (552 -> 75 lines, 17 ms)

This is a refactor. The matcher and the diff-scoped logic are carried over
unchanged, so the gate reports the same findings with the same attribution and
the same limitations. Equivalence is verified rather than assumed: the old and
new matchers agree file-by-file across every in-scope file, identical finding
sets, nothing only-in-old and nothing only-in-new.

Two behaviour notes, both called out in the issue rather than buried:

The ceiling becomes `enforce: 'info'`. The gate table requires a whole-repo count
that is not a hard zero to be informational, because one branch's inherited debt
must not fail another branch's build. The two diff-scoped checks still stop any
new site at zero tolerance.

The diff-scoped markers print only when a base ref resolved. Printing them at
zero with no base would claim a check that never ran, and would match the table's
`find` regex and suppress its own NOT RUN branch.
